WIP update
Pretty much in Skeletal / mockup the Silver key side is functional.

The rest of the map WIP.

i'm planning to add 2 secret counters one for a LG  secret and one for the doors to open with a key... THIS particular update is due to a possible bug i found (to "open" the door you must press the func_door that is the button , triggering 2 would open this door.